# Conflicts:
#	.github/pull_request_template.md
#	build.gradle
#	hhplus-api/src/main/java/com/hanghe/application/PaymentFacade.java
#	hhplus-api/src/main/java/com/hanghe/application/ReservationFacade.java
#	hhplus-api/src/main/java/com/hanghe/common/configuration/WebConfig.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/entity/Concert.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/entity/ConcertSchedule.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/entity/ConcertSeat.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/entity/SeatStatus.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/repository/ConcertScheduleRepository.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/repository/ConcertSeatRepository.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/service/ConcertSeatService.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/service/ConcertService.java
#	hhplus-api/src/main/java/com/hanghe/domain/concert/service/dto/ConcertSeatDTO.java
#	hhplus-api/src/main/java/com/hanghe/domain/payment/entity/Payment.java
#	hhplus-api/src/main/java/com/hanghe/domain/payment/repository/PaymentRepository.java
#	hhplus-api/src/main/java/com/hanghe/domain/payment/service/PaymentService.java
#	hhplus-api/src/main/java/com/hanghe/domain/payment/service/dto/PaymentChargeDTO.java
#	hhplus-api/src/main/java/com/hanghe/domain/payment/service/dto/PaymentUseDTO.java
#	hhplus-api/src/main/java/com/hanghe/domain/queue/entity/QueueToken.java
#	hhplus-api/src/main/java/com/hanghe/domain/queue/repository/QueueTokenRepository.java
#	hhplus-api/src/main/java/com/hanghe/domain/queue/service/QueueTokenService.java
#	hhplus-api/src/main/java/com/hanghe/domain/reservation/entity/Reservation.java
#	hhplus-api/src/main/java/com/hanghe/domain/reservation/repository/ReservationRepository.java
#	hhplus-api/src/main/java/com/hanghe/domain/reservation/service/ReservationService.java
#	hhplus-api/src/main/java/com/hanghe/domain/user/entity/User.java
#	hhplus-api/src/main/java/com/hanghe/domain/user/repository/UserRepository.java
#	hhplus-api/src/main/java/com/hanghe/domain/user/service/UserService.java
#	hhplus-api/src/main/java/com/hanghe/domain/user/service/dto/UserBalanceDTO.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/concert/controller/ConcertController.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/concert/dto/request/ConcertDateRequest.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/concert/dto/request/ConcertSeatRequest.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/concert/dto/response/ConcertDateResponse.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/concert/dto/response/ConcertSeatResponse.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/payment/controller/PaymentController.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/payment/dto/request/PaymentChargeRequest.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/payment/dto/request/PaymentUseRequest.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/payment/dto/response/PaymentChargeResponse.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/payment/dto/response/PaymentUseResponse.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/queueToken/scheduler/QueueTokenScheduler.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/user/controller/UserController.java
#	hhplus-api/src/main/java/com/hanghe/interfaces/user/dto/response/UserBalanceResponse.java
#	hhplus-api/src/main/resources/application.yml
#	hhplus-api/src/test/java/com/hanghe/controller/ConcertControllerTest.java
#	hhplus-api/src/test/java/com/hanghe/controller/PaymentControllerTest.java
#	hhplus-api/src/test/java/com/hanghe/controller/UserControllerTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/concert/ConcertSeatServiceTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/concert/ConcertServiceTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/payment/PaymentServiceTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/payment/PaymentTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/queue/QueueTokenServiceTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/queue/QueueTokenTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/reservation/ReservationServiceTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/reservation/ReservationTest.java
#	hhplus-api/src/test/java/com/hanghe/domain/user/UserServiceTest.java

+++ hhplus-api/src/main/java/com/hanghe/domain/payment/service/dto/PaymentChargeDTO.java
@@ -2,7 +2,7 @@ package com.hanghe.domain.payment.service.dto;
 
 import com.hanghe.domain.payment.entity.Payment;
 
-public record PaymentChargeDTO(Long paymentId, int amount) {
+public record PaymentChargeDTO(Long paymentId, Long amount) {
     public static PaymentChargeDTO from(Payment payment) {
         return new PaymentChargeDTO(
                 payment.getId(),
